// src/interface/resumen.interface.ts
// ugts = Unidades de Gestión Territorial
export interface Ugt {
	id: string;
	ugt: string;
	distritos: Distrito[];
}

export interface Distrito {
	id: string;
	nombre: string;
	coords: string;
}

// interface de los componentes de la pagina RESUMEN
export interface ComponentListProps {
	title: string;
	subtitle?: string;
	value: number | string;
	children?: React.ReactNode;
	decimalPlaces?: number;
}

export interface DistritoPotencial {
	id: string;
	nombre: string;
	coords: string;
	potencialidades: Potencialidades[];
}

export interface Potencialidades {
	id: string;
	nombre: string;
	icono: string;
	hexColor: string;
}

// interface de la data RESUMEN

export interface ResumenData {
	idh: number;
	ingreso_per_capita: number;
	brechas: Brecha[];
	brecha_financiera: number;
	proyectos_cartera: ProyectosCartera;
	recursos: Recursos;
	inversionSocial: InversionSocial;
}

export interface Brecha {
	variable: string;
	porcentaje: number;
}

export interface InversionSocial {
	inversion_menor: number;
	inversion_mayor: number;
}

export interface ProyectosCartera {
	total: Plazo;
	corto_plazo: Plazo;
	mediano_plazo: Plazo;
	largo_plazo: Plazo;
}

export interface Plazo {
	numero: number;
	monto: number;
}

export interface Recursos {
	regalia_menor: number;
	regalia_media: number;
	regalia_mayor: number;
}

// Generated by https://quicktype.io

export interface PeriodoData {
	id: string;
	titulo: string;
	value: string;
}

// Generated by https://quicktype.io